requirements.txt
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
retrying
python-dotenv
pyTelegramBotAPI
